refactor(query): Move the legacy rewrite into its own file
Two thirds of filter.go was the rewrite of a filter into the legacy
predicate fields, which is a transitional concern: it exists until each
backend evaluates the filter itself, and it will shrink and go away as the
per-backend routing lands. filter.go now holds what stays — the errors the
API layers map, the choice between passing the filter on and rewriting it,
and the capability check — and filter_legacy.go holds the rewrite, with the
tests split the same way.

No logic changed. legacyFilterLevels and the two error helpers that only
the rewrite uses moved with it; isBoolean and the two errors both files
raise stayed behind.
